+++ src/main/java/com/github/kory33/signvote/command/subcommand/VoteCommandExecutor.java
@@ -9,9 +9,9 @@ import com.github.kory33.signvote.exception.ScoreCountLimitReachedException;
 import com.github.kory33.signvote.exception.VotePointAlreadyVotedException;
 import com.github.kory33.signvote.exception.VoteSessionClosedException;
 import com.github.kory33.signvote.manager.VoteSessionManager;
-import com.github.kory33.signvote.model.Limit;
-import com.github.kory33.signvote.model.VotePoint;
-import com.github.kory33.signvote.model.VoteScore;
+import com.github.kory33.signvote.vote.Limit;
+import com.github.kory33.signvote.vote.VotePoint;
+import com.github.kory33.signvote.vote.VoteScore;
 import com.github.kory33.signvote.session.VoteSession;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
