+++ src/main/java/com/github/kory33/signvote/listeners/SignListener.java
@@ -7,7 +7,7 @@ import com.github.kory33.signvote.constants.PermissionNodes;
 import com.github.kory33.signvote.constants.SignTexts;
 import com.github.kory33.signvote.core.SignVote;
 import com.github.kory33.signvote.manager.VoteSessionManager;
-import com.github.kory33.signvote.model.VotePoint;
+import com.github.kory33.signvote.vote.VotePoint;
 import com.github.kory33.signvote.session.VoteSession;
 import org.bukkit.block.Block;
 import org.bukkit.block.Sign;
