+++ src/main/java/com/github/kory33/signvote/manager/VoteManager.java
@@ -3,9 +3,9 @@ package com.github.kory33.signvote.manager;
 import com.github.kory33.signvote.constants.Patterns;
 import com.github.kory33.signvote.exception.VotePointAlreadyVotedException;
 import com.github.kory33.signvote.exception.VotePointNotVotedException;
-import com.github.kory33.signvote.model.Vote;
-import com.github.kory33.signvote.model.VotePoint;
-import com.github.kory33.signvote.model.VoteScore;
+import com.github.kory33.signvote.vote.Vote;
+import com.github.kory33.signvote.vote.VotePoint;
+import com.github.kory33.signvote.vote.VoteScore;
 import com.github.kory33.signvote.session.VoteSession;
 import com.github.kory33.signvote.utils.FileUtils;
 import com.google.gson.Gson;
