+++ src/main/java/com/github/kory33/signvote/session/VoteSession.java
@@ -5,9 +5,9 @@ import com.github.kory33.signvote.constants.*;
 import com.github.kory33.signvote.exception.*;
 import com.github.kory33.signvote.manager.VoteLimitManager;
 import com.github.kory33.signvote.manager.VoteManager;
-import com.github.kory33.signvote.model.Limit;
-import com.github.kory33.signvote.model.VotePoint;
-import com.github.kory33.signvote.model.VoteScore;
+import com.github.kory33.signvote.vote.Limit;
+import com.github.kory33.signvote.vote.VotePoint;
+import com.github.kory33.signvote.vote.VoteScore;
 import com.github.kory33.signvote.utils.FileUtils;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
