+++ src/main/java/com/github/kory33/signvote/ui/player/AddScoreInterface.java
@@ -8,9 +8,9 @@ import com.github.kory33.signvote.configurable.JSONConfiguration;
 import com.github.kory33.signvote.constants.MagicNumbers;
 import com.github.kory33.signvote.constants.MessageConfigNodes;
 import com.github.kory33.signvote.constants.PermissionNodes;
-import com.github.kory33.signvote.model.Limit;
-import com.github.kory33.signvote.model.VoteLimit;
-import com.github.kory33.signvote.model.VoteScore;
+import com.github.kory33.signvote.vote.Limit;
+import com.github.kory33.signvote.vote.VoteLimit;
+import com.github.kory33.signvote.vote.VoteScore;
 import com.github.kory33.signvote.session.VoteSession;
 import com.github.kory33.signvote.ui.player.defaults.IDefaultFormInterface;
 import com.github.ucchyocean.messaging.tellraw.MessageParts;
