+++ src/main/java/com/github/kory33/signvote/ui/player/stats/TotalVoteStats.java
@@ -8,7 +8,7 @@ import org.bukkit.entity.Player;
 import com.github.kory33.signvote.configurable.JSONConfiguration;
 import com.github.kory33.signvote.constants.StatsType;
 import com.github.kory33.chatgui.manager.PlayerInteractiveInterfaceManager;
-import com.github.kory33.signvote.model.VotePointStats;
+import com.github.kory33.signvote.vote.VotePointStats;
 import com.github.kory33.signvote.session.VoteSession;
 import com.github.kory33.chatgui.model.player.BrowseablePageInterface;
